Modyfikacja o zadania z lekcji, float => double, XML => LINQ, Convert,(type) => TryParse

+++ Laboratorium2/MyFormatter.cs
@@ -9,12 +9,12 @@ namespace Laboratorium2
 {
     public class MyFormatter
     {
-        public static string FormatUsdPrice(double price)
+        public static string FormatUsdPrice(decimal price)
         {
             var usc = new CultureInfo("en-us");
             return price.ToString("C2", usc);
         }
-        public static string FormatPlnPrice(double price)
+        public static string FormatPlnPrice(decimal price)
         {
             var plc = new CultureInfo("pl-PL");
             return price.ToString("C2", plc);
